pyproject.toml
[project]
name = "andyterm"
version = "0.1.0"
description = "Serial & SFTP GUI tool for industrial embedded computers"
authors = [{ name = "Andy", email = "andy@moxa.com" }]
requires-python = ">=3.11"
dependencies = [
    "PySide6>=6.7",
    "pyserial>=3.5",
    "paramiko>=3.4",
    "asyncssh>=2.15",
    "pyte>=0.8",
    "qasync>=0.27",
    "keyring>=25.0",
    "cryptography>=42.0",
    "pydantic>=2.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-qt>=4.4",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.12",
    "pytest-cov>=5.0",
    "ruff>=0.5",
    "mypy>=1.10",
    "types-paramiko",
]
build = [
    "pyinstaller>=6.6",
]

[project.scripts]
andyterm = "andyterm.__main__:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E", "F", "W",     # pycodestyle + pyflakes
    "I",                # isort
    "N",                # pep8-naming
    "UP",               # pyupgrade
    "B",                # bugbear
    "SIM",              # simplify
    "RUF",              # ruff-specific
    "ASYNC",            # async lints
    "S",                # bandit (security)
]
ignore = [
    "S101",             # allow assert in tests
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S", "N"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unreachable = true
warn_redundant_casts = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = ["pyte.*", "qasync.*", "serial.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "hil: requires real Moxa hardware (skipped by default in CI)",
    "slow: slow tests",
]

[tool.coverage.run]
source = ["src/andyterm"]
omit = ["*/tests/*", "*/__main__.py"]
